[Attestation] Added runtime parameteres to the measurement
